requirements.txt
asgiref==3.4.1
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.8
django-cors-headers==3.10.0
django-decouple==2.1
django-jazzmin==2.4.8
django-oauth-toolkit==1.5.0
django-oidc-provider==0.7.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
future==0.18.2
idna==3.3
jwcrypto==1.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycparser==2.21
pycryptodomex==3.11.0
pyjwkest==1.4.2
PyJWT==2.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
